Imagenet to cuda images

+++ engine_pretrain.py
@@ -30,7 +30,7 @@ def train_one_epoch(model: torch.nn.Module, data_loader: Iterable, optimizer: to
         with torch.cuda.amp.autocast():
             if args.use_decs:
                 # breakpoint()
-                samples.cuda()
+                samples.cuda().half()
                 loss_comb, _, _,_,_,_ = model(samples)
                 loss_mage, loss_mage_spot = loss_comb
                 # breakpoint()
